+++ src/app/jobs/(jobs-category)/public-jobs/page.tsx
@@ -1,16 +1,19 @@
 "use client";
-
 import SelectedChips from "@/features/jobs/components/SelectedChips";
 import { useEffect } from "react";
-import JobCard from "../../../features/home/components/JobCard";
-import JobFilter from "../../../features/jobs/components/JobFilter";
-import { useFilterTabStore } from "../../../stores/useJobFilterStore";
+import JobCard from "../../../../features/home/components/JobCard";
+import JobFilter from "../../../../features/jobs/components/JobFilter";
+import { useFilterTabStore } from "../../../../stores/useJobFilterStore";
 
-export default function JobsByFieldPage() {
+export default function PublicJobsPage() {
+  const setShowLocation = useFilterTabStore((state) => state.setShowLocation);
   const setShowJobs = useFilterTabStore((state) => state.setShowJobs);
+  const setShowOtherConditions = useFilterTabStore((state) => state.setShowOtherConditions);
 
   useEffect(() => {
-    setShowJobs(true);
+    setShowLocation(false);
+    setShowJobs(false);
+    setShowOtherConditions(false);
   }, []);
 
   return (
@@ -22,7 +25,7 @@ export default function JobsByFieldPage() {
       <div className="bg-gray-z-light py-6">
         <section className="w-full max-w-7xl mx-auto my-8 px-4 ">
           <div className="flex justify-between items-center py-6 mb-4">
-            <h2 className="text-2xl font-semibold">직종별 공고</h2>
+            <h2 className="text-2xl font-semibold">공공일자리</h2>
           </div>
           <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-7">
             <JobCard />
